pyproject.toml
[project]
name = "azure-document-intelligence-vs-markitdown-vs-tika"
version = "0.1.0"
description = ""
authors = [
    {name = "kimtth"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "azure-ai-documentintelligence (>=1.0.2,<2.0.0)",
    "tika (>=3.1.0,<4.0.0)",
    "markitdown[pdf] (>=0.1.2,<0.2.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "xlsxwriter (>=3.2.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
